Add files via upload
In the selection retransmission protocol, the sender needs to maintain a timer for each unconfirmed data packet to ensure that it can be retransmitted when the data packet times out and no confirmation is received.
Only one global timer is used in the current code, which cannot meet the requirement of timing each unconfirmed data packet separately.
I can't modify the emulator.c code. I can only send the start timer using the first package, so I need to roll back.